requirements.txt
alembic==1.9.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
blinker==1.5
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
flake8==6.0.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.3
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
polyline==2.0.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
python-dotenv==0.21.1
requests==2.28.2
SQLAlchemy==2.0.0
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
WTForms==3.0.1
